# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Sys API server.",
        "title": "Sys API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/reports": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves paginated and filtered interview reports for the authenticated candidate.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "List past interview reports",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by title or summary",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by difficulty (beginner, intermediate, senior, staff)",
                        "name": "difficulty",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"data\": models.ReportsListResponse}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"error\": \"unauthorized\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/reports/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves multi-axis rubrics, strengths, weaknesses, and AI remarks for a session ID. Returns 202 Accepted if evaluation is in progress.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reports"
                ],
                "summary": "Get detailed evaluation report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Interview Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"data\": models.DetailedReportResponse}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "202": {
                        "description": "{\"status\": \"evaluating\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"error\": \"unauthorized\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "{\"success\": false, \"error\": \"report not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves workspace, role target, AI interviewer persona, and canvas grid configuration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get candidate settings and preferences",
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"data\": models.UserSettings}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"error\": \"unauthorized\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upserts workspace preferences, target companies, AI persona strictness, and canvas defaults.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update candidate settings and preferences",
                "parameters": [
                    {
                        "description": "Settings update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MonuChaudhary14_Archon_internal_models.UserSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"message\": \"Settings updated successfully\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"error\": \"unauthorized\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"success\": false, \"error\": \"...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MonuChaudhary14_Archon_internal_models.UserSettings": {
            "type": "object",
            "properties": {
                "auto_save_interval_seconds": {
                    "type": "integer"
                },
                "canvas_grid_type": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_notifications": {
                    "type": "boolean"
                },
                "enable_proactive_hints": {
                    "type": "boolean"
                },
                "enable_voice_interview": {
                    "type": "boolean"
                },
                "export_format": {
                    "type": "string"
                },
                "feedback_style": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "interviewer_strictness": {
                    "type": "string"
                },
                "primary_stack": {
                    "type": "string"
                },
                "snap_to_grid": {
                    "type": "boolean"
                },
                "target_companies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "target_level": {
                    "type": "string"
                },
                "weekly_interview_target": {
                    "type": "integer"
                },
                "weekly_report_digest": {
                    "type": "boolean"
                },
                "years_of_experience": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}